HeroesOOP/src/main/java/efs.task.oop/ExtraordinaryVillager.java
package efs.task.oop;

public class ExtraordinaryVillager extends Villager {
    Skill skill;

    public ExtraordinaryVillager(String name, int age, Skill skill) {
        super(name, age);
        this.skill = skill;
    }

    public enum Skill {
        IDENTIFY("I will identify items for you at no charge."),
        SHELTER("I can offer you poor shelter.");

        private String text;

        Skill(String text) {
            this.text = text;
        }
        public String getText() {
            return text;
        }
    }

    @Override
    public void sayHello() {
        System.out.println("Greetings traveler... I'm " + name + " and I'm " + age + " years old. " + skill.getText());
    }

    @Override
    public void attack(Fighter victim) {
        victim.takeHit(0);
    }

    @Override
    public void takeHit(int damage){
        health = 0;
    }
}
